(feat) extract reusable logo component and refactor nav links
